+++ src/main/java/org/tlesis/squakeplusplus/scheduler/ClientTickHandler.java
@@ -1,6 +1,5 @@
 package org.tlesis.squakeplusplus.scheduler;
 
-import org.tlesis.squakeplusplus.event.RenderHandler;
 
 import fi.dy.masa.malilib.interfaces.IClientTickHandler;
 import net.minecraft.client.MinecraftClient;
@@ -13,7 +12,6 @@ public class ClientTickHandler implements IClientTickHandler {
 
         if (mc.world != null && mc.player != null) {
             TaskScheduler.getInstanceClient().runTasks();
-            RenderHandler.renderHandler.updateData(mc);
         }
         
         if (mc.player != null) {
